app-backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4
aiosignal==1.3.1 ; python_version >= '3.7'
alembic==1.10.4 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
blinker==1.6.2 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
flask==2.3.1
flask-cors==3.0.10
flask-migrate==4.0.4
flask-sqlalchemy==3.0.3
frozenlist==1.3.3 ; python_version >= '3.7'
gevent==22.10.2
greenlet==2.0.2 ; platform_python_implementation == 'CPython'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.6.0 ; python_version < '3.10'
importlib-resources==5.12.0 ; python_version < '3.9'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
psutil==5.9.5
pymysql==1.0.3
python-dotenv==1.0.0
requests==2.29.0
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.11 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.2 ; python_version >= '3.8'
yarl==1.9.2 ; python_version >= '3.7'
zipp==3.15.0 ; python_version >= '3.7'
zope.event==4.6
zope.interface==6.0 ; python_version >= '3.7'
